AM62: Secure Boot: provide SIGN_KEYDIR helper class
When building for Secure Boot, the variable SIGN_KEYDIR must be present
and it must point to some existing path.

When this variable is not set (or not pointing to a directory), there
are some hard-to-understand error messages. This commit ensures that
there are proper error messages for these error cases.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>