am62x-var-som-secureboot: provide machines for Secure Boot variant
This allows to build images / packages, specifically for Secure Boot.
This provides a straight-forward way for Yocto recipes to make
modifications when building for secured devices, checking for the
MACHINEOVERRIDES 'var-som-secureboot'.

This also links to distinctive U-Boot defconfig files so that the
resultung binaries of the U-Boot will match the requirements for the
Yocto build system.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>