bcm-utils: update for Scarthgap
This provides the updates necessary to make this run on Scarthgap.
This includes an update of the syntax for gpioset.

Also use extra parameter for faster firmware upload to achieve the same
BT uptime as before.

Also access device symlinks to correct UART / I2C devices instead of
hard-coded numbers.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>