bcm-utils: fix race condition for I2C EEPROM
BCM-Utils requires to be able to read from the I2C EEPROM. In order to
this,

- the kernel must have probed the I2C driver
- the udev rule for this I2C driver must have been processed

When the BCM-Utils script has been executed before both have happened,
the WiFi device will not come up.
There is a timeout of 10 seconds to wait for /dev/i2c-var-eeprom. After
this timeout, it will not be known if there should be a WiFi device and
WiFi bringup will fail.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>